+++ R/ndraws.R
@@ -2,9 +2,7 @@
 #'
 #' @export
 #' @export ndraws
-#' @rdname ndraws-dynamitefit
 #' @aliases ndraws
-#' @method ndraws dynamitefit
 #' @param x \[`dynamitefit`]\cr The model fit object.
 #' @return Number of posterior draws as a single `integer` value.
 #' @examples